+++ R/tvAR.R
@@ -76,9 +76,9 @@
 #' predictability with new estimators of realized variance and variance 
 #' risk premium. Url= http://pure.au.dk/portal/files/123066669/rp18_10.pdf
 #'
-#' Chen, X. B., Gao, J., Li, D., and Silvapulle, P (2017) Nonparametric estimation and 
-#' forecasting for time-varying coefficient realized volatility models,
-#' \emph{Journal of Business \& Economic Statistics}, online, 1-13.
+#' Chen, X. B., Gao, J., Li, D., and Silvapulle, P (2017) Nonparametric Estimation and 
+#' Forecasting for Time-Varying Coefficient Realized Volatility Models.
+#' \emph{Journal of Business and Economic Statistics}, 36, 88-100.
 #' 
 #' Corsi, F. (2009) A simple approximate long-memory model of realized 
 #' volatility. \emph{Journal of Financial Econometrics}, 7, 174-196.