+++ R/tvOLS.R
@@ -18,7 +18,7 @@ tvOLS <- function(x, ...) UseMethod("tvOLS", x)
 #' @param y A vector with dependent variable.
 #' @param z A vector with the variable over which coefficients are smooth over.
 #' @param ez (optional) A scalar or vector with the smoothing values. If 
-#' values are included then the vector z is used.
+#' values are not included then the vector \code{z} is used instead.
 #' @param bw A numeric vector.
 #' @inheritParams tvLM
 #' @param singular.ok	Logical. If FALSE, a singular model is an error.