+++ R/tvRE.R
@@ -19,7 +19,7 @@ tvRE <- function(x, ...) UseMethod("tvRE", x)
 #' @param y A vector with dependent variable.
 #' @param z A vector with the variable over which coefficients are smooth over.
 #' @param ez (optional) A scalar or vector with the smoothing values. If 
-#' values are included then the vector z is used.
+#' values are not included then the vector \code{z} is used instead.
 #' @param bw A numeric vector with the bandwidth.
 #' @param Sigma NULL (default) or a matrix of size obs x obs..
 #' @param neq A scalar with the number of equations